+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Construction of `MockEventQueue` now includes a constructor argument for the time-fetching function
 * Construction of `PinHistory` now includes a constructor argument for the time-fetching function
 * `PinHistory` can now return an array of timestamps for its events
+* `GodmodeState` is now a singleton pattern, which is necessary to support the globality of Arduino functions
+* `GodmodeState` now uses timestamped PinHistory for Analog and Digital
 
 ### Deprecated